Changed most of the parameters to psdsim() into doubles. Changed some betad[i][j] to betd (since i hasn't been initialized at that point. Continued porting main Fortran a little past the first call to psdsim()

+++ trunk/blzsim.h
@@ -72,8 +72,8 @@ class BlzSim {
   // N=number of data points in the lc, should be an integer power of 2, N<=8192
   // t_incre1=increment in time in each step while resampling the simulated data.
   // This must be larger than the smallest interval between successive data points in the input light curve
-  void psdsim(const int N, const float beta1, const float beta2, const float nu_break, 
-              const float t_incre1, float *lc_sim);
+  void psdsim(const int N, const double beta1, const double beta2, const double nu_break, 
+              const double t_incre1, double *lc_sim);
 
   double seedph(const double f);
   double ajnu(const double anu);
